backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
httpx==0.28.1
pydantic==2.10.5
pydantic-settings==2.7.1
PyMuPDF==1.25.3
Pillow==11.1.0
prometheus-client==0.21.1
python-jose[cryptography]==3.3.0
aiosqlite==0.20.0
sqlalchemy[asyncio]==2.0.36
alembic==1.14.1
openai==1.59.9
